+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.0.2
 
 * Change tpl tags from { ... } to {{ ... }}
 * update _compile()
+* update namespace from Antonraharja to Playsms
 * Change dynamic vars to allow only ready variables for an echo
 
 Version 1.0.1